secunik/frontend_streamlit/requirements.txt
streamlit==1.29.0
streamlit-option-menu==0.3.6
streamlit-ace==0.1.1
streamlit-aggrid==0.3.4.post3
plotly==5.17.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
python-magic==0.4.27
watchdog==3.0.0